+++ crates/biome_js_analyze/src/lint/nursery.rs
@@ -23,6 +23,7 @@ pub mod use_focusable_interactive;
 pub mod use_import_extensions;
 pub mod use_import_restrictions;
 pub mod use_number_to_fixed_digits_argument;
+pub mod use_semantic_elements;
 pub mod use_sorted_classes;
 pub mod use_throw_new_error;
 pub mod use_top_level_regex;
@@ -52,6 +53,7 @@ declare_group! {
             self :: use_import_extensions :: UseImportExtensions ,
             self :: use_import_restrictions :: UseImportRestrictions ,
             self :: use_number_to_fixed_digits_argument :: UseNumberToFixedDigitsArgument ,
+            self :: use_semantic_elements :: UseSemanticElements ,
             self :: use_sorted_classes :: UseSortedClasses ,
             self :: use_throw_new_error :: UseThrowNewError ,
             self :: use_top_level_regex :: UseTopLevelRegex ,