+++ packages/@biomejs/backend-jsonrpc/src/workspace.ts
@@ -1086,7 +1086,7 @@ export interface Nursery {
 	 */
 	useGenericFontNames?: RuleConfiguration_for_Null;
 	/**
-	 * Require import extensions for relative imports.
+	 * Enforce file extensions for relative imports.
 	 */
 	useImportExtensions?: RuleFixConfiguration_for_Null;
 	/**
@@ -1097,6 +1097,10 @@ export interface Nursery {
 	 * Enforce using the digits argument with Number#toFixed().
 	 */
 	useNumberToFixedDigitsArgument?: RuleFixConfiguration_for_Null;
+	/**
+	 * It detects the use of role attributes in JSX elements and suggests using semantic elements instead.
+	 */
+	useSemanticElements?: RuleConfiguration_for_Null;
 	/**
 	 * Enforce the sorting of CSS utility classes.
 	 */
@@ -2307,6 +2311,7 @@ export type Category =
 	| "lint/nursery/useImportExtensions"
 	| "lint/nursery/useImportRestrictions"
 	| "lint/nursery/useNumberToFixedDigitsArgument"
+	| "lint/nursery/useSemanticElements"
 	| "lint/nursery/useSortedClasses"
 	| "lint/nursery/useThrowNewError"
 	| "lint/nursery/useTopLevelRegex"